backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
requests==2.32.3
stripe==11.1.0
slowapi==0.1.9
fastapi-mail==1.5.1
google-generativeai==0.8.3
email-validator==2.2.0
alembic==1.14.0

